Move interaction param to new file
